Extract BioMart queries into a separate module
